+++ libs/elodin-editor/src/ui/schematic/tree.rs
@@ -99,6 +99,7 @@ impl WidgetSystem for TreeWidget<'_, '_> {
                         }
                     }
                     impeller2_wkt::SchematicElem::Line3d(_line_3d) => {}
+                    impeller2_wkt::SchematicElem::PointTrails(_trails) => {}
                     impeller2_wkt::SchematicElem::VectorArrow(_arrow) => {}
                     impeller2_wkt::SchematicElem::WorldMesh(_world_mesh) => {}
                     impeller2_wkt::SchematicElem::Window(_window) => {}